Import live.2004.07.06.tar.gz
- Modified the code so that it will build for WinCE.
  (Thanks to Gabriel Bouvigne for these changes.)

+++ liveMedia/MP3StreamState.cpp
@@ -386,7 +386,7 @@ unsigned MP3StreamState::readFromStream(unsigned char* buf,
 
     return totBytesRead;
   } else {
-    waitUntilSocketIsReadable(fEnv, fileno(fFid));
+    waitUntilSocketIsReadable(fEnv, (int)fileno(fFid));
     return fread(buf, 1, numChars, fFid);
   }
 }
